requirements.txt
aim==3.29.1
azureml-core==1.58.0
azureml-fsspec==1.3.1
azure-monitor-opentelemetry==1.6.8
azure-keyvault-secrets==4.7.0
azure-ai-ml==1.21.1
azure-identity==1.19.0
mltable==1.6.1
numpy==1.26.4
pandas==2.2.3
pre-commit==3.8.0
python-dotenv==1.0.1
randomname==0.2.1
scikit_learn==1.6.1
omegaconf==2.3
setuptools==78.1.1
matplotlib==3.10.0
xgboost==3.0.0
marshmallow==3.23.2
pymsteams==0.2.2
ydata-profiling==4.18.1